Bugfixing of memory whole and improved printout

+++ cpu_code.h
@@ -1,7 +1,7 @@
 #pragma once
 
 void tryKeyRandom(int i, char *nonce, char*veribuf);
-void measureCPUPerformance(unsigned long nrOfThreads,
-						   unsigned long *nrKeysCalculatedResult,
-						   unsigned long *nrOfSecondsInTotalMeasured,
-						   unsigned long nrSecondsToMeasure = 30);
+void measureCPUPerformance(unsigned long long nrOfThreads,
+						   unsigned long long *nrKeysCalculatedResult,
+						   unsigned long long *nrOfSecondsInTotalMeasured,
+						   unsigned long long nrSecondsToMeasure = 30);
