+++ gpu_code.h
@@ -5,7 +5,7 @@ void tryKeysGPUSingleShot(unsigned int nrBlocks,
 				uint8_t nonce_hc[8],
 		        char *verificationBuffer,
 				char *keys,
-				unsigned long nrKeys,
+				unsigned long long nrKeys,
 				bool *result);
 
 void tryKeysGPUMultiShot(unsigned int nrBlocks,
@@ -13,13 +13,13 @@ void tryKeysGPUMultiShot(unsigned int nrBlocks,
 				uint8_t nonce_hc[8],
 		        char *verificationBuffer,
 				char*keys,
-				unsigned long nrKeys,
-				unsigned long keysBeforeContextSwitch,
-				unsigned long keysInTotalToCalculate);
+				unsigned long long nrKeys,
+				unsigned long long keysBeforeContextSwitch,
+				unsigned long long keysInTotalToCalculate);
 
 void measureGPUPerformance(unsigned int nrBlocks,
 		        unsigned int nrThreads,
-				unsigned long keysBeforeContextSwitch,
-				unsigned long *nrKeysCalculatedResult,
-				unsigned long *nrOfSecondsInTotalMeasured,
+				unsigned long long keysBeforeContextSwitch,
+				unsigned long long *nrKeysCalculatedResult,
+				unsigned long long *nrOfSecondsInTotalMeasured,
 				int nrSecondsToMeasure = 30);
